v1.0.5

+++ src/main/java/com/github/microprograms/micro_api_sdk/utils/ApiDeployUtils.java
@@ -28,8 +28,12 @@ public class ApiDeployUtils {
     private static void deploySh(EngineDefinition engineDefinition) throws IOException {
         String home = engineDefinition.getDeployDefinition().getRemoteJavaApplicationHome();
         remoteSsh(String.format("mkdir -p %s/bin", home), engineDefinition);
-        remoteSsh(String.format("echo -e \"%s\" > %s/bin/start.sh", _buildStartSh(engineDefinition).replaceAll("\"", "\\\\\"").replaceAll("\\$!", "\\\\\\$!"), home), engineDefinition);
-        remoteSsh(String.format("echo -e \"%s\" > %s/bin/stop.sh", _buildStopSh(engineDefinition).replaceAll("`", "\\\\`"), home), engineDefinition);
+        remoteSsh(String.format("echo -e \"%s\" > %s/bin/start.sh", _escapeEcho(_buildStartSh(engineDefinition)), home), engineDefinition);
+        remoteSsh(String.format("echo -e \"%s\" > %s/bin/stop.sh", _escapeEcho(_buildStopSh(engineDefinition)), home), engineDefinition);
+    }
+
+    private static String _escapeEcho(String cmd) throws IOException {
+        return cmd.replaceAll("\"", "\\\\\"").replaceAll("\n", "\\\\n").replaceAll("\\$!", "\\\\\\$!").replaceAll("`", "\\\\`");
     }
 
     private static String _buildStartSh(EngineDefinition engineDefinition) throws IOException {
@@ -56,11 +60,15 @@ public class ApiDeployUtils {
     }
 
     private static void deployLib(EngineDefinition engineDefinition) throws IOException {
+        String isExistSetenvSh = localSsh("[[ -f ~/.micro_api_sdk/setenv.sh ]]; echo $?;", engineDefinition);
+        if (Integer.valueOf(isExistSetenvSh.replaceAll("\\s*", "")) != 0) {
+            throw new RuntimeException("canot found \"~/.micro_api_sdk/setenv.sh\" file, Maven may not work properly.");
+        }
         String localHome = engineDefinition.getDeployDefinition().getLocalMavenProjectHome();
         String remoteHome = engineDefinition.getDeployDefinition().getRemoteJavaApplicationHome();
         String remoteHost = engineDefinition.getServerAddressDefinition().getHost();
         String cd = String.format("cd %s;", localHome);
-        String setenv = "sh ~/.micro_api_sdk/setenv.sh;";
+        String setenv = "source ~/.micro_api_sdk/setenv.sh;";
         String jar = "mvn clean package; mvn dependency:copy-dependencies -DoutputDirectory=lib; cp target/*.jar lib/;";
         String rsync = String.format("export RSYNC_PASSWORD=pass; rsync -vzrtopg --delete --progress lib/ common@%s::common/%s/lib/;", remoteHost, remoteHome);
         String del = "rm -rf lib/;";
@@ -83,17 +91,19 @@ public class ApiDeployUtils {
         remoteSsh(String.format("sudo cat %s/server.nohup.out", home), engineDefinition);
     }
 
-    public static void remoteSsh(String cmd, EngineDefinition engineDefinition) throws IOException {
+    public static String remoteSsh(String cmd, EngineDefinition engineDefinition) throws IOException {
         DeployDefinition deployDefinition = engineDefinition.getDeployDefinition();
         Shell shell = new SshByPassword(engineDefinition.getServerAddressDefinition().getHost(), deployDefinition.getRemoteSshPort(), deployDefinition.getRemoteSshUser(), deployDefinition.getRemoteSshPassword());
         String stdout = new Shell.Plain(shell).exec(cmd);
-        log.info("remoteSsh> " + cmd + "\n" + stdout);
+        log.info("remoteSsh> {}{}", cmd, StringUtils.isBlank(stdout) ? "" : "\n" + stdout);
+        return stdout;
     }
 
-    public static void localSsh(String cmd, EngineDefinition engineDefinition) throws IOException {
+    public static String localSsh(String cmd, EngineDefinition engineDefinition) throws IOException {
         DeployDefinition deployDefinition = engineDefinition.getDeployDefinition();
         Shell shell = new SshByPassword("localhost", deployDefinition.getLocalSshPort(), deployDefinition.getLocalSshUser(), deployDefinition.getLocalSshPassword());
         String stdout = new Shell.Plain(shell).exec(cmd);
-        log.info("localSsh> " + cmd + "\n" + stdout);
+        log.info("localSsh> {}{}", cmd, StringUtils.isBlank(stdout) ? "" : "\n" + stdout);
+        return stdout;
     }
 }
